IO Tested and works
I changed all the c getlines to istream getlines so that I could do
interactive debugging since c getlines require the input to be
immediately available and do not wait.
Also, once you get an int with the istream, it still leaves behind the
newline after it so I made sure to read that too so it would all be
reading the right things.