Two Changes
Not initializing int in loop to 0 was making me see weird things.
I think you meant to subtract one from the first pick since the
candidate array starts at 0. Eg: first pick of first ballot is 1, that
is for John Doe, which is index 0 of the candidate array. Correct me if
I’m wrong.